+++ src/main/java/com/ness/movie_release_web/model/wrapper/tmdb/company/search/CompanySearchWrapper.java
@@ -18,9 +18,9 @@ public class CompanySearchWrapper {
     @JsonProperty("page")
     private Integer page;
     @JsonProperty("total_results")
-    private Integer totalResults;
+    private Long totalResults;
     @JsonProperty("total_pages")
-    private Integer totalPages;
+    private Long totalPages;
     @JsonProperty("results")
     private List<ProductionCompanyWrapper> results = new ArrayList<>();
 }
