+++ src/main/java/com/ness/movie_release_web/service/SubscriptionService.java
@@ -1,8 +1,8 @@
 package com.ness.movie_release_web.service;
 
 public interface SubscriptionService {
-    void subscribeToMovie(Integer tmdbId, String login);
-    void unsubscribeFromMovie(Integer tmdbId, String login);
-    void subscribeToSeries(Integer tmdbId, String login);
-    void unsubscribeFromSeries(Integer tmdbId, String login);
+    void subscribeToMovie(Long tmdbId, String login);
+    void unsubscribeFromMovie(Long tmdbId, String login);
+    void subscribeToSeries(Long tmdbId, String login);
+    void unsubscribeFromSeries(Long tmdbId, String login);
 }
